create app.js, update server.spec.js
